feat: l'amelioration de la structure du code pour la gestion des benevoles et pour la gestion des stands. L'amelioration du design.
