+++ src/components/gestions/GestionCategoriesStands.tsx
@@ -1,4 +1,5 @@
 import {
+    Box,
     Button, Dialog, DialogActions, DialogContent, DialogTitle,
     Paper,
     Table,
@@ -20,6 +21,7 @@ import {DeleteCategory} from "../../services_REST/serveur/admin/categories/Delet
 import {useStandsStore} from "../../stores/StandsStore.ts";
 import {GetStands} from "../../services_REST/serveur/admin/stands/GetStands.ts";
 import {SuccessMessage} from "../SuccessMessage.tsx";
+import {styleCustomInput} from "../../styles/CustomInputField.ts";
 
 export const GestionCategories = () => {
     const {categories, addCategorie, updateCategorie, deleteCategorie} = useCategoriesStore();
@@ -33,26 +35,6 @@ export const GestionCategories = () => {
     const [open, setOpen] = useState(false);
     const [isEditing, setIsEditing] = useState(false);
     const [formData, setFormData] = useState<Categorie>(new Categorie(0, ""));
-
-    const styleCustom = {
-        '& label.Mui-focused': {
-            color: '#2C2C2C',
-        },
-        '& .MuiInput-underline:after': {
-            borderBottomColor: '#7f5656',
-        },
-        '& .MuiOutlinedInput-root': {
-            '& fieldset': {
-                borderColor: '#7f5656',
-            },
-            '&:hover fieldset': {
-                borderColor: '#7f5656',
-            },
-            '&.Mui-focused fieldset': {
-                borderColor: '#7f5656',
-            },
-        }
-    }
 
     const handleOpen = (editing = false, categorie: Categorie | null = null) => {
         setIsEditing(editing);
@@ -143,16 +125,18 @@ export const GestionCategories = () => {
 
     return (
         <>
-            <div style={{ padding: "20px", textAlign: "center" }}>
+            <Box sx={{ p: 3, textAlign: "center" }}>
                 <Typography variant="h5" sx={{ mt: 1 }}>
                     Gestion des catégories de stands
                 </Typography>
                 <Button variant="contained" onClick={() => handleOpen(false)} style={{ marginTop: "20px", backgroundColor: "#7f5656" }}>
                     Ajouter une catégorie
                 </Button>
-            </div>
+            </Box>
+
             <SuccessMessage successMessage={successMessage} setSuccessMessage={setSuccessMessage}/>
-            <div style={{ padding: "20px", display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
+
+            <Box sx={{ p: 3, mb: 15, display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
                 <TableContainer component={Paper} sx={{maxHeight: 250, maxWidth: 600, boxShadow: 4, overflow: "auto", borderRadius: 2}}>
                     <Table sx={{ border: "1px solid #ddd" }}>
                         <TableHead>
@@ -182,12 +166,12 @@ export const GestionCategories = () => {
                         </TableBody>
                     </Table>
                 </TableContainer>
-            </div>
+            </Box>
 
             <Dialog open={open} onClose={handleClose}>
                 <DialogTitle>{isEditing ? "Modifier la catégorie" : "Ajouter une catégorie"}</DialogTitle>
                 <DialogContent>
-                    <TextField fullWidth margin="dense" variant="outlined" sx={styleCustom} label="Nom" name="nom_categorie" value={formData.nom_categorie} onChange={handleChange} error={!!nomError} helperText={nomError}/>
+                    <TextField fullWidth margin="dense" variant="outlined" sx={styleCustomInput} label="Nom" name="nom_categorie" value={formData.nom_categorie} onChange={handleChange} error={!!nomError} helperText={nomError}/>
                     {error && (
                         <Typography color="error" variant="body2" sx={{ mt: 1 }}>
                             {error}
