+++ src/pages/administration/GestionFestivaliers.tsx
@@ -1,6 +1,7 @@
 import {Header} from "../../components/Header.tsx";
 import {Footer} from "../../components/Footer.tsx";
 import {
+    Box,
     Button,
     Dialog, DialogActions, DialogContent,
     DialogTitle, FormControl,
@@ -34,6 +35,7 @@ import {useBenevolesStore} from "../../stores/BenevolesStore.ts";
 import {updateFestivaliers} from "../../services/festivaliers.ts";
 import {SnackbarError} from "../../components/SnackbarError.tsx";
 import {SuccessMessage} from "../../components/SuccessMessage.tsx";
+import {styleCustomInput} from "../../styles/CustomInputField.ts";
 
 export const GestionFestivaliers = () => {
     const {festivaliers, setFestivaliers, addFestivalier, updateFestivalier, deleteFestivalier} = useFestivaliersStore();
@@ -75,26 +77,6 @@ export const GestionFestivaliers = () => {
             return () => clearTimeout(timer);
         }
     }, [successMessage]);
-
-    const styleCustom = {
-        '& label.Mui-focused': {
-            color: '#2C2C2C',
-        },
-        '& .MuiInput-underline:after': {
-            borderBottomColor: '#7f5656',
-        },
-        '& .MuiOutlinedInput-root': {
-            '& fieldset': {
-                borderColor: '#7f5656',
-            },
-            '&:hover fieldset': {
-                borderColor: '#7f5656',
-            },
-            '&.Mui-focused fieldset': {
-                borderColor: '#7f5656',
-            },
-        }
-    }
 
     const handleOpen = (editing = false, user: User | null = null) => {
         setIsEditing(editing);
@@ -230,19 +212,18 @@ export const GestionFestivaliers = () => {
     return (
         <>
             <Header />
-            <div style={{height: "1100px"}}>
-            <div style={{ padding: "20px", textAlign: "center" }}>
+            <Box sx={{ p: 3, textAlign: "center" }}>
                 <Typography variant="h5" sx={{ mt: 1 }}>
                     Gestion des festivaliers
                 </Typography>
                 <Button variant="contained" onClick={() => handleOpen(false)} style={{ marginTop: "20px", backgroundColor: "#7f5656" }}>
                     Ajouter un festivalier
                 </Button>
-            </div>
+            </Box>
 
             <SuccessMessage successMessage={successMessage} setSuccessMessage={setSuccessMessage}/>
 
-            <div style={{ padding: "20px" }}>
+            <Box sx={{ p: 3, mb: 15 }}>
                 <TableContainer component={Paper} sx={{maxHeight: 400, boxShadow: 4, overflow: "auto", borderRadius: 2}}>
                     <Table sx={{ border: "1px solid #ddd" }}>
                         <TableHead>
@@ -277,15 +258,15 @@ export const GestionFestivaliers = () => {
                         </TableBody>
                     </Table>
                 </TableContainer>
-            </div>
+            </Box>
 
             <Dialog open={open} onClose={handleClose}>
                 <DialogTitle>{isEditing ? "Modifier le festivalier" : "Ajouter un festivalier"}</DialogTitle>
                 <DialogContent>
-                    <TextField fullWidth margin="dense" variant="outlined" sx={styleCustom} label="Nom" name="nom" value={formData.nom} onChange={handleChange} error={!!errors.nom} helperText={errors.nom}/>
-                    <TextField fullWidth margin="dense" variant="outlined" sx={styleCustom} label="Prénom" name="prenom" value={formData.prenom} onChange={handleChange} error={!!errors.prenom} helperText={errors.prenom}/>
-                    <TextField fullWidth margin="dense" variant="outlined" sx={styleCustom} label="Login" name="login" value={formData.login} onChange={handleChange} error={!!errors.username} helperText={errors.username}/>
-                    {!isEditing && (<TextField fullWidth margin="dense" variant="outlined" sx={styleCustom} label="Mot de passe" name="password" type="password" value={password} onChange={handlePasswordChange} error={!!errors.password} helperText={errors.password}/>)}
+                    <TextField fullWidth margin="dense" variant="outlined" sx={styleCustomInput} label="Nom" name="nom" value={formData.nom} onChange={handleChange} error={!!errors.nom} helperText={errors.nom}/>
+                    <TextField fullWidth margin="dense" variant="outlined" sx={styleCustomInput} label="Prénom" name="prenom" value={formData.prenom} onChange={handleChange} error={!!errors.prenom} helperText={errors.prenom}/>
+                    <TextField fullWidth margin="dense" variant="outlined" sx={styleCustomInput} label="Login" name="login" value={formData.login} onChange={handleChange} error={!!errors.username} helperText={errors.username}/>
+                    {!isEditing && (<TextField fullWidth margin="dense" variant="outlined" sx={styleCustomInput} label="Mot de passe" name="password" type="password" value={password} onChange={handlePasswordChange} error={!!errors.password} helperText={errors.password}/>)}
                     {error && (
                         <Typography color="error" variant="body2" sx={{ mt: 1 }}>
                             {error}
@@ -301,7 +282,7 @@ export const GestionFestivaliers = () => {
             <Dialog open={openRoleDialog} onClose={handleCloseRoleDialog}>
                 <DialogTitle>Changer le rôle</DialogTitle>
                 <DialogContent>
-                    <FormControl fullWidth margin="dense" sx={styleCustom} error={!!roleError}>
+                    <FormControl fullWidth margin="dense" sx={styleCustomInput} error={!!roleError}>
                         <InputLabel id="role-label">Rôle</InputLabel>
                         <Select
                             label="Rôle"
@@ -329,9 +310,10 @@ export const GestionFestivaliers = () => {
                     <Button onClick={handleRoleSubmit} sx={{backgroundColor: "#7f5656"}} variant="contained">Modifier</Button>
                 </DialogActions>
             </Dialog>
+
             <SnackbarError error={errorSnackbar} setError={setSnackbarError}/>
-            </div>
-            <Footer />
+
+            <Footer/>
         </>
     )
 }
