+++ src/pages/administration/GestionTerminaux.tsx
@@ -1,6 +1,7 @@
 import {Header} from "../../components/Header.tsx";
 import {Footer} from "../../components/Footer.tsx";
 import {
+    Box,
     Button,
     Paper,
     Table,
@@ -61,14 +62,13 @@ export const GestionTerminaux = () => {
     return (
         <>
             <Header/>
-            <div style={{height: "1065px"}}>
-                <div style={{padding: "20px", textAlign: "center" }}>
+                <Box sx={{p: 3, textAlign: "center" }}>
                     <Typography variant="h5" sx={{ mt: 1 }}>
                         Gestion des terminaux
                     </Typography>
-                </div>
+                </Box>
                 <SuccessMessage successMessage={successMessage} setSuccessMessage={setSuccessMessage}/>
-                <div style={{ padding: "20px" }}>
+                <Box sx={{ p: 3, mb: 15 }}>
                     <TableContainer component={Paper} sx={{maxHeight: 400, boxShadow: 4, overflow: "auto", borderRadius: 2}}>
                         <Table sx={{ border: "1px solid #ddd" }}>
                             <TableHead>
@@ -99,9 +99,8 @@ export const GestionTerminaux = () => {
                             </TableBody>
                         </Table>
                     </TableContainer>
-                </div>
+                </Box>
                 <SnackbarError error={errorSnackbar} setError={setSnackbarError}/>
-            </div>
             <Footer/>
         </>
     )
