+++ src/services_REST/serveur/connection/Login.ts
@@ -8,7 +8,7 @@ const URL_AUTH = parameters.URL_AUTH
 export const Login = async(username: string, password: string): Promise<Token | null> => {
     try {
         const response = await axios.post<Token>(`${URL_SERVER}${URL_AUTH}`, {username, password}, {withCredentials: true});
-        console.log(`response.data de requette TokenLWT ${response.data.token}`)
+        console.log(`response.data de requette Login ${response.data.token}`)
         return response.data
     } catch (error) {
         if (error instanceof AxiosError) {
